css/css-values/subresource-criticality/tentative/font-face-unused.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>unused @font-face blocks load</title>
<link rel=help href="https://github.com/w3c/csswg-drafts/issues/1088">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
@font-face {
  font-family: Revalia;
  src: url("/fonts/Revalia.woff?pipe=trickle(d1)");
}
</style>
<script>
var t1 = performance.timing.responseStart;

async_test(function(t) {
window.addEventListener("load", t.step_func_done(function() {
    var t2 = Date.now();
    var diff = t2 - t1;
    assert_less_than(diff, 1000, "ms after style");
  }));
}, "load doesn't wait for unused @font-face");
</script>
